+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 1.5.x - 2024xxxx Maintenance release for 1.x (PHP >= 7.4)
   * ...
 
+3.x.x - 2024xxxx
+  * Update packages + check for vendor directory
+
 3.2.0 - 20240905 Update tests + upgrade code
   * Update phpunit tests + upgrade code with rector